pos to long

+++ src/main/java/io/kaitai/struct/ByteBufferKaitaiStream.java
@@ -162,7 +162,7 @@ public class ByteBufferKaitaiStream extends KaitaiStream {
     }
 
     @Override
-    public int pos() {
+    public long pos() {
         return bb.position();
     }
